Crate nue_io

Utilities for working with I/O streams.

Reexports

pub use read_exact::ReadExactExt;

Modules

read_exact
An extension for Reading an exact amount of data.

Structs

BufSeeker
A buffered reader that allows for seeking within the buffer.

ReadWriteTell
A wrapper that implements Tell for streams that don't support it.

Region
Creates an isolated segment of an underlying stream.

SeekAbsoluteRewind
An absolute seeking wrapper around a Tell + SeekForward + SeekRewind type.

SeekAll
A wrapper that decomposes Seek into its individual traits.

SeekForwardRead
A forward seeking wrapper around a Read type.

SeekForwardWrite
A forward seeking wrapper around a Write type.

Take
Wraps around a stream to limit the length of the underlying stream.

Traits

SeekAbsolute
A limited form of seeking that only seeks to an absolute position from the start.

SeekAlignExt
An extension trait that will seek to meet a specified alignment.

SeekBackward
A limited form of seeking that only allows seeking backward.

SeekEnd
A limited form of seeking that only seeks to an offset from the end of the stream.

SeekForward
A limited form of seeking that only allows seeking forward.

SeekRewind
A limited form of seeking that can only be reset from the beginning.

Tell
Exposes the current position in the stream without changing it.