Module nue_io::seek_forward

Extensions for seeking around streams that do not normally support it.

Structs

SeekForwardRead
A forward seeking wrapper around a Read type.

SeekForwardRewind
A forward seeking wrapper around a Read + SeekRewind type.

SeekForwardWrite
A forward seeking wrapper around a Write type.

Traits

SeekForward
Extension trait that wraps a stream in a new forward seeker.

SeekRewind
A limited form of seeking that can only be reset from the beginning.