Trait nue_io::SeekEnd

pub trait SeekEnd {
    fn seek_end(&mut self, offset: i64) -> Result<u64>;
}

A limited form of seeking that only seeks to an offset from the end of the stream.

Required Methods

fn seek_end(&mut self, offset: i64) -> Result<u64>

Seeks to the end of the stream + offset.

Returns the new position in the stream.